+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/AbstractIn1804DeviceSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804
+{
+	public abstract class AbstractIn1804DeviceSettings : AbstractExtronSwitcherDeviceSettings, IIn1804DeviceSettings
+	{
+	}
+}