+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/IIn1804DeviceSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804
+{
+	public interface IIn1804DeviceSettings : IExtronSwitcherDeviceSettings
+	{
+	}
+}