+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804/In1804Device.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804
+{
+	public sealed class In1804Device : AbstractIn1804Device<In1804DeviceSettings>
+	{
+	}
+}