+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804/In1804DeviceSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804
+{
+	[KrangSettings("In1804", typeof(In1804Device))]
+	public sealed class In1804DeviceSettings : AbstractIn1804DeviceSettings
+	{
+	}
+}