+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804Di/In1804DiDevice.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804Di
+{
+	public sealed class In1804DiDevice : AbstractIn1804Device<In1804DiDeviceSettings>
+	{
+	}
+}