+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804Di/In1804DiDeviceSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804Di
+{
+	[KrangSettings("In1804Di", typeof(In1804DiDevice))]
+	public sealed class In1804DiDeviceSettings : AbstractIn1804DeviceSettings
+	{
+	}
+}