+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804DiDo/In1804DiDoDevice.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804DiDo
+{
+	public sealed class In1804DiDoDevice : AbstractIn1804Device<In1804DiDoDeviceSettings>
+	{
+	}
+}