+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804DiDo/In1804DiDoDeviceSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804DiDo
+{
+	[KrangSettings("In1804DiDo", typeof(In1804DiDoDevice))]
+	public sealed class In1804DiDoDeviceSettings : AbstractIn1804DeviceSettings
+	{
+	}
+}