+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804Do/In1804DoDevice.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804Do
+{
+	public sealed class In1804DoDevice : AbstractIn1804Device<In1804DoDeviceSettings>
+	{
+	}
+}