+++ ICD.Connect.Routing.Extron/Devices/Switchers/In1804/In1804Do/In1804DoDeviceSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.Extron.Devices.Switchers.In1804.In1804Do
+{
+	[KrangSettings("In1804Do", typeof(In1804DoDevice))]
+	public sealed class In1804DoDeviceSettings : AbstractIn1804DeviceSettings
+	{
+	}
+}